+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Alert user of unpaid invoices during rollover. Refs UIF-378.
 * Add test rollover action to rollover settings. Refs UIF-393.
 * Fiscal year rollover test confirmation modal. Refs UIF-395.
+* Export list of unpaid invoices. Refs UIF-394.
 
 ## [3.1.1](https://github.com/folio-org/ui-finance/tree/v3.1.1) (2022-03-31)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.1.0...v3.1.1)